fix: suppress cast_possible_truncation clippy lint in embedding pooling

+++ mistralrs-core/src/embedding_models/layers.rs
@@ -47,7 +47,8 @@ impl Module for Pooling {
         if self.pooling_mode_mean_tokens {
             // Assume full attention mask. Otherwise this must be updated
             // Mean pooling: sum(embeddings) / seq_len
-            let seq_len = xs.dim(1)? as u32 as f64;
+            #[allow(clippy::cast_possible_truncation)]
+            let seq_len = xs.dim(1)? as f64;
             let mean_embeddings = (xs.sum(1)? / seq_len)?;
             outputs.push(mean_embeddings);
         }
